chore: update pest to 2.x & allow for comments within enums & structs

+++ omg-idl-gen/files/test-vectors/typedef_array_dim_1/expected.rs
@@ -1,4 +1,4 @@
 
 #[allow(dead_code)]
 #[allow(non_camel_case_types)]
-pub type Foo = [u8;2];
+pub type Foo = [u8;2 as usize];
