+++ omg-idl-gen/files/test-vectors/typedef_array_dim_2/expected.rs
@@ -1,4 +1,4 @@
 
 #[allow(dead_code)]
 #[allow(non_camel_case_types)]
-pub type Foo = [[u8;2];3+3];
+pub type Foo = [[u8;2 as usize];3+3 as usize];
